split text processing method to  2 functions
